backend/requirements.txt
flask
flask-cors
numpy
tensorflow-cpu
pillow
gunicorn
gdown
